+++ database/migrations/2019_12_23_105358_create_order_items_table.php
@@ -20,7 +20,7 @@ class CreateOrderItemsTable extends Migration
             $table->foreignId('order_id');
             $table->foreignId('product_id');
             $table->string('name');
-            $table->integer('price');
+            $table->decimal('price');
             $table->integer('quantity');
             $table->timestamps();
         });
